Create Api module, add error message, add Loading component
