+++ src/styles.css
@@ -48,7 +48,9 @@ body {
   font-weight: 700;
   font-family: 'Kanit', sans-serif;
   font-variant: small-caps;
-  font-size: 25px;
+  font-size: 15px;
+  color: white;
+  background-color: rgb(127, 172, 255);
 }
 
 /* DEGREE BUTTON */
@@ -128,4 +130,31 @@ input:checked + .slider::after {
 
 .slider.round::before {
   border-radius: 50%;
+}
+
+/* LOADING  */
+
+#loading {
+  width: 3rem;
+  height: 3rem;
+  border: 6px solid #f3f3f3;
+  border-top: 6px solid hsl(226, 83%, 61%);
+  border-radius: 100%;
+  margin: auto;
+  visibility: hidden;
+  animation: spin 1.5s infinite linear;
+}
+
+#loading.display {
+  visibility: visible;
+}
+
+@keyframes spin {
+  from {
+    transform: rotate(0deg);
+  }
+
+  to {
+    transform: rotate(360deg);
+  }
 }
